requirements.txt
Flask 
pyodbc 
python-dotenv
datetime
requests
gunicorn